@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 class Background extends StatelessWidget {
   final Widget child;
-  const Background({Key key, @required this.child}) : super(key: key);
+  const Background({Key? key, required this.child}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
