@@ -11,24 +11,39 @@ class StudentRegisterScreen extends StatefulWidget {
 class _StudentRegisterScreenState extends State<StudentRegisterScreen> {
   final _form = GlobalKey<FormState>();
 
-  final Map<Student, Student> _formData = {};
-
-  void _getFormData(Student student) {
-    if (student != null) {
-      student.id as Student;
-      student.name as Student;
-      student.email as Student;
-      student.attendance as Student;
-      student.avatarUrl as Student;
-    }
-  }
+  final _formData = Map<String, Object>();
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
-    final Student student = ModalRoute.of(context).settings.arguments;
 
-    _getFormData(student);
+    if (_formData.isEmpty) {
+      final student = ModalRoute.of(context)!.settings.arguments as Student;
+
+      // ignore: unnecessary_null_comparison
+      if (student != null) {
+        _formData['id'] = student.id;
+        _formData['name'] = student.name;
+        _formData['gender'] = student.gender;
+        _formData['email'] = student.email;
+        _formData['attendance'] = student.attendance;
+        _formData['avatarUrl'] = student.avatarUrl;
+        _formData['belt'] = student.belt;
+        _formData['degree'] = student.degree;
+        _formData['birthdate'] = student.birthdate;
+
+        //_imageUrlController.text = _formData['imageUrl'];
+      } else {
+        _formData['name'] = '';
+        _formData['gender'] = 0;
+        _formData['email'] = "";
+        _formData['attendance'] = 0;
+        _formData['avatarUrl'] = '';
+        _formData['belt'] = 0;
+        _formData['degree'] = 0;
+        _formData['birthdate'] = DateTime.now();
+      }
+    }
   }
 
   @override
@@ -47,18 +62,22 @@ class _StudentRegisterScreenState extends State<StudentRegisterScreen> {
               color: Colors.white,
             ),
             onPressed: () {
-              final isValid = _form.currentState.validate();
+              final isValid = _form.currentState!.validate();
               if (isValid) {
-                _form.currentState.save();
-                // Provider.of<Students>(context, listen: false).put(
-                //   Student(
-                //     id: _formData['id'],
-                //     name: _formData['name'],
-                //     email: _formData['email'],
-                //     attendance: _formData['attendance'],
-                //     avatarUrl: _formData['avatarUrl'],
-                //   ),
-                //);
+                _form.currentState!.save();
+                Provider.of<Students>(context, listen: false).put(
+                  Student(
+                    id: _formData['id'] as int,
+                    name: _formData['name'] as String,
+                    email: _formData['email'] as String,
+                    attendance: _formData['attendance'] as int,
+                    avatarUrl: _formData['avatarUrl'] as String,
+                    belt: _formData['belt'] as int,
+                    birthdate: _formData['birthdate'] as DateTime,
+                    degree: _formData['degree'] as int,
+                    gender: _formData['gender'] as int,
+                  ),
+                );
                 Navigator.of(context).pop();
               }
             },
@@ -74,13 +93,13 @@ class _StudentRegisterScreenState extends State<StudentRegisterScreen> {
               TextFormField(
                 initialValue: _formData.isEmpty
                     ? "Cadastrar novo aluno"
-                    : "Matrícula", // + _formData['id'],
+                    : "Matrícula " + _formData['id'].toString(),
                 decoration: InputDecoration(
                   enabled: false,
                 ),
               ),
               TextFormField(
-                initialValue: "MUDAR",
+                initialValue: _formData['name'].toString(),
                 decoration: InputDecoration(
                   labelText: 'Nome',
                   icon: Icon(Icons.person),
@@ -97,7 +116,7 @@ class _StudentRegisterScreenState extends State<StudentRegisterScreen> {
                 //onSaved: (value) => _formData['name'] = value as Student,
               ),
               TextFormField(
-                // initialValue: _formData['email'],
+                initialValue: _formData['email'].toString(),
                 decoration: InputDecoration(
                   labelText: 'E-mail',
                   icon: Icon(Icons.mail),
@@ -114,7 +133,7 @@ class _StudentRegisterScreenState extends State<StudentRegisterScreen> {
                 //onSaved: (value) => _formData['email'] = value as Student,
               ),
               TextFormField(
-                // initialValue: _formData['attendance'],
+                initialValue: _formData['attendance'].toString(),
                 decoration: InputDecoration(
                   labelText: 'Presenças',
                   icon: Icon(Icons.date_range_rounded),
@@ -122,7 +141,47 @@ class _StudentRegisterScreenState extends State<StudentRegisterScreen> {
                 //onSaved: (value) => _formData['attendance'] = value as Student,
               ),
               TextFormField(
-                //initialValue: _formData['avatarUrl'],
+                initialValue: _formData['gender'].toString(),
+                decoration: InputDecoration(
+                  labelText: 'Gênero',
+                  icon: Icon(Icons.help_outline),
+                ),
+                //onSaved: (value) => _formData['avatarUrl'] = value as Student,
+              ),
+              TextFormField(
+                initialValue: _formData['belt'].toString(),
+                decoration: InputDecoration(
+                  labelText: 'Faixa',
+                  icon: Icon(Icons.minimize),
+                ),
+                //onSaved: (value) => _formData['avatarUrl'] = value as Student,
+              ),
+              TextFormField(
+                initialValue: _formData['degree'].toString(),
+                decoration: InputDecoration(
+                  labelText: 'Grau',
+                  icon: Icon(Icons.view_column),
+                ),
+                //onSaved: (value) => _formData['avatarUrl'] = value as Student,
+              ),
+              TextFormField(
+                initialValue: _formData['birthdate'].toString(),
+                decoration: InputDecoration(
+                  labelText: 'Data de inicio',
+                  icon: Icon(Icons.celebration),
+                ),
+                //onSaved: (value) => _formData['avatarUrl'] = value as Student,
+              ),
+              TextFormField(
+                initialValue: _formData['birthdate'].toString(),
+                decoration: InputDecoration(
+                  labelText: 'Data de nascimento',
+                  icon: Icon(Icons.cake),
+                ),
+                //onSaved: (value) => _formData['avatarUrl'] = value as Student,
+              ),
+              TextFormField(
+                initialValue: _formData['avatarUrl'].toString(),
                 decoration: InputDecoration(
                   labelText: 'Foto',
                   icon: Icon(Icons.image_search),
