@@ -1,8 +1,7 @@
 class Constants {
-  static const BASE_API_URL = 'http://localhost:44324/api';
+  static const BASE_API_URL = '192.168.1.49:44324';
 
-  static const STUDENT_VERSION_1 = 'v1/student';
-  static const ADDRESS_VERSION_1 = 'v1/address';
-  static const PERSON_VERSION_1 = 'v1/person';
-
+  static const STUDENT_VERSION_1 = '/api/v1/Student';
+  static const ADDRESS_VERSION_1 = '/api/v1/Address';
+  static const PERSON_VERSION_1 = '/api/v1/Person';
 }
